refactor: merge system/game/ + gameplay/ + root orphans into ui/game/
Move 28 source files into apps/web/src/components/ui/game/, create a
single index.ts barrel, fix all internal relative imports, update all
external consumers, delete system/game/ and gameplay/ directories and
the GameplayHud.tsx shim. TypeScript build passes with zero errors.
