--- apps/web/src/components/GameplayHud.tsx
@@ -1,17 +1,0 @@
-export {
-  BoardStage,
-  FloatingPlayerRoster,
-  GameOverLayer,
-  InspectorDock,
-  NotificationPipe,
-  PlayerRoster,
-  PrivateHandRail,
-  SupplyDock,
-  TicketChoiceSheet,
-  TicketDock,
-  TurnIndicator,
-  formatCardLabel,
-  type GameplayNotification,
-  type PlayerRosterEntry,
-  type PlayerRosterTimer
-} from "./gameplay";
