--- apps/web/src/components/gameplay/index.ts
@@ -1,7 +1,0 @@
-export { BoardStage } from "./BoardStage";
-export { InspectorDock } from "./InspectorDock";
-export { GameOverLayer, NotificationPipe } from "./notifications";
-export { PrivateHandRail, TicketChoiceSheet, TicketDock } from "./PlayerHandPanel";
-export { FloatingPlayerRoster, PlayerRoster, type PlayerRosterEntry, type PlayerRosterTimer } from "./PlayerRosterPanel";
-export { SupplyDock } from "./SupplyDock";
-export { TurnIndicator, formatCardLabel, type GameplayNotification } from "../system/game";
