+++ apps/web/src/components/local/EndgameGrid.tsx
@@ -1,5 +1,5 @@
 import { type GameState, type MapConfig } from "@hudson-hustle/game-core";
-import { EndgameBreakdown } from "../EndgameBreakdown";
+import { EndgameBreakdown } from "../ui/game";
 import { SurfaceCard } from "../ui/primitives/SurfaceCard";
 
 interface EndgameGridProps {
