+++ apps/web/src/components/local/RouteBuildPanel.tsx
@@ -2,7 +2,7 @@ import type { MapConfig, TrainCardColor } from "@hudson-hustle/game-core";
 import { getCityName } from "@hudson-hustle/game-core";
 import { ChoiceChipButton } from "../ui/primitives/ChoiceChipButton";
 import { SurfaceCard } from "../ui/primitives/SurfaceCard";
-import { formatCardLabel } from "../GameplayHud";
+import { formatCardLabel } from "../ui/game";
 
 interface RouteDef {
   id: string;
