--- apps/web/src/components/system/game/index.ts
@@ -1,12 +1,0 @@
-export { CardSlot, formatCardLabel } from "./CardSlot";
-export type { CardSlotProps } from "./CardSlot";
-export { GameOverPanel } from "./GameOverPanel";
-export { NotificationStack } from "./NotificationStack";
-export type { GameplayNotification } from "./NotificationStack";
-export { SeatTile } from "./SeatTile";
-export type { SeatTileProps } from "./SeatTile";
-export { SideTabRail } from "./SideTabRail";
-export type { SideTab } from "./SideTabRail";
-export { TicketSlip } from "./TicketSlip";
-export type { TicketSlipStatus } from "./TicketSlip";
-export { TurnIndicator } from "./TurnIndicator";
