+++ apps/web/src/components/ui/game/BoardStage.stories.tsx
