+++ apps/web/src/components/ui/game/GameOverPanel.stories.tsx
