+++ apps/web/src/components/ui/game/InspectorDock.stories.tsx
