+++ apps/web/src/components/ui/game/SeatTile.stories.tsx
