+++ apps/web/src/components/ui/game/SeatTile.tsx
