+++ apps/web/src/components/ui/game/SideTabRail.stories.tsx
