+++ apps/web/src/components/ui/game/SideTabRail.tsx
