+++ apps/web/src/components/ui/game/TicketSlip.stories.tsx
