+++ apps/web/src/components/ui/game/TicketSlip.tsx
