+++ apps/web/src/components/ui/game/TransitCard.tsx
